+++ worker/src/mails_api/send_mail_api.ts
@@ -256,7 +256,7 @@ api.post('/api/send_mail', async (c) => {
         await sendMail(c, address, reqJson);
     } catch (e) {
         console.error("Failed to send mail", e);
-        return c.text(`Failed to send mail ${(e as Error).message}`, 400)
+        return c.text("Failed to send mail", 400)
     }
     return c.json({ status: "ok" })
 })
@@ -274,7 +274,7 @@ api.post('/external/api/send_mail', async (c) => {
         return c.json({ status: "ok" })
     } catch (e) {
         console.error("Failed to send mail", e);
-        return c.text(`Failed to send mail ${(e as Error).message}`, 400)
+        return c.text("Failed to send mail", 400)
     }
 })